# Documentation/Sprint 15.11.2017 - 22.11.2017/SwaggerDocumentation.yaml
swagger: '2.0'
info:
  description: >
    This is the Swagger Documentation for the Interaction API. For more
    information please contact either member of Team Interaction.
  version: 1.0.0
  title: Interaction API
...
paths:
...
  /users:
    get:
      tags:
        - Users
      summary: Retrieve all users
      responses:
        '200':
          description: Successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/User'
    post:
      tags:
        - Users
      summary: Create new user
      parameters:
        - in: body
          name: body
          description: Configuration for user that should be created
          required: true
          schema:
            $ref: '#/definitions/User'
      responses:
        '200':
          description: Successful operation
  '/users/{id}':
    delete:
      tags:
        - Users
      summary: Delete user by id
      responses:
        '200':
          description: Successful operation
    get:
      tags:
        - Users
      summary: Retrieve user by id
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/User'
    put:
      tags:
        - Users
      summary: Update user by id
      parameters:
        - in: body
          name: body
          description: Updated Values
          required: true
          schema:
            $ref: '#/definitions/User'
      responses:
        default:
          description: successful operation
    parameters:
      - in: path
        name: id
        type: integer
        required: true
        description: User id
definitions:
  Address:
    type: object
    properties:
      country:
        type: string
      city:
        type: string
      street:
        type: string
      houseNr:
        type: integer
...
  PaymentOption:
    type: object
    properties:
      provider:
        type: string
        description: A PaymentOption that the user configured for the account
        enum:
          - PayPal
          - Credit Card
...
  User:
    type: object
    properties:
      id:
        type: integer
        description: Unique identifier for that user
      firstName:
        type: string
      lastName:
        type: string
      email:
        type: string
      password:
        type: string
      address:
        $ref: '#/definitions/Address'
      paymentOptions:
        type: array
        items:
          $ref: '#/definitions/PaymentOption'
      userType:
        type: string
        description: The role that the user has in our system
        enum:
          - Customer
          - Factory Owner